Remove last domain participant services
