+++ dds/src/implementation/domain_participant_backend/mod.rs
@@ -2,4 +2,3 @@ pub mod domain_participant_actor;
 pub mod domain_participant_actor_mail;
 pub mod entities;
 pub mod handle;
-pub mod services;
