--- dds/src/implementation/domain_participant_backend/services/mod.rs
@@ -1,2 +1,0 @@
-pub mod discovery_service;
-pub mod message_service;
